feat: update import paths for DTO classes in test files and refactor token service usage
